backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy>=2.0
asyncpg
pydantic>=2.0
pydantic-settings
httpx
python-jose[cryptography]
supabase
structlog
pandas
langchain
langchain-openai
langchain-community
langfuse
reportlab
razorpay
